Refactor project structure: separate views for workers, tasks, and positions
- Moved worker views to `worker_views.py`
- Moved task views to `task_views.py`
- Moved position views to `position_views.py`
- Created `base.py` for general views
- Updated `urls.py` to reflect new view locations
- Cleaned up imports for clarity and organization

This refactoring organizes the codebase, making it cleaner and more maintainable. By separating views into specific files, it improves readability and facilitates easier management and scaling of the project in the future.
