+++ tasks/views/position_views.py
@@ -1,0 +1,9 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
+from django.views import generic
+
+from tasks.models import Position
+
+
+class PositionListView(LoginRequiredMixin, generic.ListView):
+    model = Position
+    paginate_by = 8
